server/storage_api: update TestRangeResponse unit test, to work correctly with secondary application tenant
What was there before: the change focuses on updating the unit test which was failing with secondary tenant while hitting the
_status/range/{id} endpoint. This was happening because the the test had a hard-coded range id `1` to be
fetched from the aforementioned endpoint. And this range id lies outside the scope of a secondary application
tenant.

resolution: made changes to the test to include a range id which lies well within the scope of the secondary tenant.
Fixes: #110018
Part of: https://cockroachlabs.atlassian.net/browse/CRDB-31224
Epic: CRDB-38968

Release note: None